+++ lib/acyort.js
@@ -1,22 +1,20 @@
 const Logger = require('./logger')
 const Fetcher = require('./fetcher')
-const markeder = require('./markeder')
 const Processor = require('./processor')
-const Render = require('./render')
+const Generator = require('./generator')
 
 class Acyort {
   constructor() {
     this._logger = new Logger()
     this._fetcher = new Fetcher()
-    this._markeder = markeder
     this._processor = new Processor()
-    this._render = new Render()
+    this._generator = new Generator()
   }
 
   start() {
     this._fetcher.fetch()
     .then(data => this._processor.process(data))
-    .then(data => this._render.generate(data))
+    .then(data => this._generator.generate(data))
     .catch(err => this._logger.error(err))
   }
 }
