+++ lib/generator/helper/i18n.js
