+++ lib/generator/helper/index.js
