+++ lib/generator/rss.js
