+++ lib/generator/source.js
