+++ lib/render/template.js
@@ -1,7 +1,7 @@
 const swig = require('swig-templates')
 const path = require('path')
-const config = require('../../config')
-const Logger = require('../../logger')
+const config = require('../config')
+const Logger = require('../logger')
 
 swig.setDefaults({ cache: false })
