+++ lib/server/index.js
@@ -6,9 +6,8 @@ const Events = require('./events')
 const config = require('../config')
 const Logger = require('../logger')
 const Fetcher = require('../fetcher')
-const markeder = require('../markeder')
 const Processor = require('../processor')
-const Render = require('../render')
+const Generator = require('../generator')
 
 class Ws extends StaticServer {
   constructor() {
@@ -16,10 +15,9 @@ class Ws extends StaticServer {
 
     this._config = config
     this._logger = new Logger()
-    this._markeder = markeder
     this._fetcher = new Fetcher()
     this._processor = new Processor()
-    this._render = new Render()
+    this._generator = new Generator()
     this._events = new Events()
 
     this._clients = []
@@ -71,14 +69,14 @@ class Ws extends StaticServer {
 
   _refresh(type) {
     if (this._data) {
-      return this._render.genHtmls(this._data, type)
+      return this._generator.genHtmls(this._data, type)
     }
 
     return this._fetcher.fetch()
     .then(data => this._processor.process(data))
     .then((data) => {
       this._data = data
-      return this._render.genHtmls(data, type)
+      return this._generator.genHtmls(data, type)
     })
   }
 
@@ -112,11 +110,11 @@ class Ws extends StaticServer {
       }
 
       if (event === 'css') {
-        this._render.copySources()
+        this._generator.copySources()
         return this._sendEvent('css')
       }
       if (event === 'static') {
-        this._render.copySources()
+        this._generator.copySources()
         return this._sendEvent('reload')
       }
       if (event === 'yml' || event === 'html') {
